core(scoring): extract existing tests into separate file
