+++ lighthouse-core/test/report/v2/report-generator-test.js
@@ -35,92 +35,6 @@ describe('ReportGeneratorV2', () => {
     });
   });
 
-  describe('#arithmeticMean', () => {
-    it('should work for empty list', () => {
-      assert.equal(ReportGeneratorV2.arithmeticMean([]), 0);
-    });
-
-    it('should work for equal weights', () => {
-      assert.equal(ReportGeneratorV2.arithmeticMean([
-        {score: 10, weight: 1},
-        {score: 20, weight: 1},
-        {score: 3, weight: 1},
-      ]), 11);
-    });
-
-    it('should work for varying weights', () => {
-      assert.equal(ReportGeneratorV2.arithmeticMean([
-        {score: 10, weight: 2},
-        {score: 0, weight: 7},
-        {score: 20, weight: 1},
-      ]), 4);
-    });
-
-    it('should work for missing values', () => {
-      assert.equal(ReportGeneratorV2.arithmeticMean([
-        {weight: 1},
-        {score: 30, weight: 1},
-        {weight: 1},
-        {score: 100},
-      ]), 10);
-    });
-  });
-
-  describe('#generateReportJson', () => {
-    it('should return a score', () => {
-      const result = new ReportGeneratorV2().generateReportJson({
-        categories: {
-          'categoryA': {weight: 1, audits: [{id: 'auditA', weight: 1}]},
-          'categoryB': {weight: 4, audits: [{id: 'auditB', weight: 1}]},
-          'categoryC': {audits: []},
-        },
-      }, {auditA: {score: 50}, auditB: {score: 100}});
-
-      assert.equal(result.score, 90);
-    });
-
-    it('should return categories', () => {
-      const result = new ReportGeneratorV2().generateReportJson({
-        categories: {
-          'my-category': {name: 'My Category', audits: []},
-          'my-other-category': {description: 'It is a nice category', audits: []},
-        },
-      }, {});
-
-      assert.equal(result.categories.length, 2);
-      assert.equal(result.categories[0].name, 'My Category');
-      assert.equal(result.categories[1].description, 'It is a nice category');
-    });
-
-    it('should score the categories', () => {
-      const auditResults = {
-        'my-audit': {rawValue: 'you passed'},
-        'my-boolean-audit': {score: true, extendedInfo: {}},
-        'my-scored-audit': {score: 100},
-        'my-failed-audit': {score: 20},
-        'my-boolean-failed-audit': {score: false},
-      };
-
-      const result = new ReportGeneratorV2().generateReportJson({
-        categories: {
-          'my-category': {audits: [{id: 'my-audit'}]},
-          'my-scored': {
-            audits: [
-              {id: 'my-boolean-audit', weight: 1},
-              {id: 'my-scored-audit', weight: 1},
-              {id: 'my-failed-audit', weight: 1},
-              {id: 'my-boolean-failed-audit', weight: 1},
-            ],
-          },
-        },
-      }, auditResults);
-
-      assert.equal(result.categories.length, 2);
-      assert.equal(result.categories[0].score, 0);
-      assert.equal(result.categories[1].score, 55);
-    });
-  });
-
   describe('#generateHtmlReport', () => {
     it('should return html', () => {
       const result = new ReportGeneratorV2().generateReportHtml({});
